0.0.4 release - fix start/stop mining events being sent when pointer unlocked
Previously, the mining start event would always be sent when the mouse button
is pressed, even if the game area is not under player control. With voxel-mine
this would cause the initial damage overlay to appear whenever you clicked,
even in dialogs (such as voxel-debug or voxel-inventory-dialog) instead of the
game. To fix this, listen for 'interact' attain/release events, and only send
our mining events if interaction is attained.